+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod cargo_test;
 pub mod output;
+pub mod test_name_formatter;
 
 use cargo_test as ct;
 use output as o;
